backend/requirements.txt
fastapi
uvicorn
python-multipart
httpx
langchain-google-genai
langchain-core
python-dotenv
pillow
langchain-openai
langchain-community
pydub
openai
websockets
yt-dlp
